move logo to assets
